% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{add_station_to_table}
\alias{add_station_to_table}
\title{Add station to upstream table}
\usage{
add_station_to_table(stations, table)
}
\arguments{
\item{stations}{a list of station names to be added to the table}

\item{table}{a data.table describing upstream / downstream relationships between
stations.}
}
\description{
Creates a new table w.  Because of the limitations of the
data.table object on which the upstream-downstream tables are based, the object
is not modified in-place. Instead, a new object must be created.  A practical
workaround is to assign the new table object the same name as the old table
}
\examples{
a <- initialize_matrix_from_stations(c('st1', 'st2','st8'))
set_upstream('st1', a, c('st2', 'st8'))
a <- add_station_to_table(c('st4', 'st5'), a)

}
