% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEM_tools.R
\name{ARCTICTileURL}
\alias{ARCTICTileURL}
\title{ARCTIC DEM Tile URL}
\usage{
ARCTICTileURL(lon, lat, name)
}
\arguments{
\item{lon}{longitude in degrees as an integer}

\item{lat}{latitude in degrees as an integer}

\item{name}{(optional) tilename}
}
\value{
file path
}
\description{
Returns the URL of a ARCTIC DEM elevation tile
}
