% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydat_extras.R
\name{bool.check}
\alias{bool.check}
\title{Create a SpatialPointsDataFrame from ECDE table export.}
\usage{
bool.check(vec)
}
\arguments{
\item{vec}{A vector (any type)}
}
\value{
the original unchanged vector, or if possible a boolean
}
\description{
Changes vectors into Boolean when appropriate
}
\details{
Convert to Boolean
}
\keyword{internal}
