% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_stations.R
\name{bool_check}
\alias{bool_check}
\title{Convert vector to logical}
\usage{
bool_check(vec)
}
\arguments{
\item{vec}{A vector (any type)}
}
\value{
the original unchanged vector, or if possible a boolean
}
\description{
Changes vectors into Boolean when appropriate
}
\keyword{internal}
