% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGA_Tools.R
\name{BreachDepressionsRS}
\alias{BreachDepressionsRS}
\title{Sink Removal}
\usage{
BreachDepressionsRS(in.DEM, out.DEM, saga.env, METHOD = 0, THRESHOLD = 0,
  THRESHOLD.HEIGHT = 100, verbose = F)
}
\arguments{
\item{in.DEM}{character path to grid to fill}

\item{out.DEM}{(optional) path to save output DEM, otherwise appends '_filled' to filename of in.DEM}

\item{saga.env}{A SAGA geoprocessing object.  Suggested version is 2.2.2.}

\item{METHOD}{Available Choices: [0] Deepen Drainage Routes [1] Fill Sinks.}

\item{THRESHOLD}{logical, whether or not to exclude deeper sinks from filling}

\item{THRESHOLD.HEIGHT}{numeric, The parameter describes the maximum depth of a sink to be considered for removal [map units].
This makes it possible to exclude deeper sinks from filling.}
}
\description{
Remove sinks by filling sinks or deepening drainage routes using the method of Conrad (2001)
}
\keyword{internal}
