% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEM_tools.R
\name{CDEDGetTilePath}
\alias{CDEDGetTilePath}
\title{Get URL of CDED NTS tile}
\usage{
CDEDGetTilePath(NTS)
}
\arguments{
\item{NTS}{character NTS string in format DDDLDD or DDDL (d=digit, L=letter)
e.g 075M05 or 014K}
}
\value{
character path to CDED DEM tile for selected NTS sheet
}
\keyword{internal}
