% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEM_tools.R
\name{CDEMGetTilePath}
\alias{CDEMGetTilePath}
\title{Get URL of CDEM NTS tile}
\usage{
CDEMGetTilePath(NTS, product = "CDEM")
}
\arguments{
\item{NTS}{character, NTS string in format DDDLDD or DDDL (D=digit, L=letter)
e.g 075M05 or 014K}

\item{product}{character string specifying which data product to get (one of
"CDEM', 'CDED', 'CDSM')}
}
\value{
character path to CDED DEM tile for selected NTS sheet
}
\keyword{internal}
