% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeonamesAPI.R
\name{CGNDBDownloadAPI}
\alias{CGNDBDownloadAPI}
\title{Download data from CGNDB}
\usage{
CGNDBDownloadAPI(dstfile, terms, format)
}
\arguments{
\item{terms}{a list() of named search terms, see details for more info}

\item{format}{Character string; one of c('json', 'html', 'csv', 'kml', 'gml', 'wkt').}
}
\value{
file path to downloaded file
Canadian Geographic Names Database
}
\description{
Download data from CGNDB
}
