% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonames_db.R
\name{CGNDBHydroKML}
\alias{CGNDBHydroKML}
\title{download kml of hydrological feature}
\usage{
CGNDBHydroKML(station, dstfile)
}
\arguments{
\item{station}{name of a HYDAT hydrometric station or the name of a
waterbody}

\item{dstfile}{character string specifying a kml file to download}
}
\value{
a Spatial* object
}
\description{
Based on the name of a HYDAT survey station, attempts to download a
kml corresponding to the hydrological feature being measured.
}
\details{
After downloading, the kml is converted to a shapefile. Because of the
limitations of feature types in a shapefile, only one geometry type is preserved
(points, lines or polygons). The script does its best to figure out
which of the three should be preserved based on the station name.

CONCISE CODES: "RIV", "LAKE", "SEA", "CHAN", "BAY",
}
