% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{check_symmetry}
\alias{check_symmetry}
\title{Check symmetry of station matrix}
\usage{
check_symmetry(table)
}
\arguments{
\item{table:}{a data table created from initialize_matrix_from_stations. An
(n x n+1) data.table with first column named "ID" and all other columns
named according to station number. Row i and column i+1 should correspond to
the same station.}
}
\value{
A list with an element for every x,y station pair that violates
the antisymmetry assumption
}
\description{
Finds all monitoring stations that violate the assumption
of antisymmetry for upstream/downstream relations.
}
\details{
The antisymmetry assumption states that if x is upstream of y then
y is not upstream of x
}
