% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{check_transitivity}
\alias{check_transitivity}
\title{Check transitivity of station matrix}
\usage{
check_transitivity(table)
}
\arguments{
\item{table}{a data table created from initialize_matrix_from_stations. An
(n x n+1) data.table with first column named "ID" and all other columns
named according to station number. Row i and column i+1 should correspond to
the same station.}
}
\value{
A list with an element for every x,y station pair that violates
the transitivity assumption or an empty vector
}
\description{
Finds all monitoring stations that violate the assumption
of transitivity for upstream/downstream relations.
}
\details{
The transitivity assumption states that if x is upstream of y and
y is upstream of z, then x is upstream of z. This tests that assumption for a
matrix
}
