% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGA_Tools.R
\name{ClipGridRS}
\alias{ClipGridRS}
\title{Clip grid to point plus error  # Clip Grids = 31}
\usage{
ClipGridRS(point, in.DEM, out.DEM, saga.env, tol, verbose = F)
}
\arguments{
\item{point}{SpatialPointsDataFrame}

\item{in.DEM}{character path to tiled grid to clip}

\item{out.DEM}{(optional) path to save output DEM, otherwise appends '_clipped' to filename of in.DEM}

\item{saga.env}{A SAGA geoprocessing object.  Suggested version is 2.2.2.}

\item{tol}{size of clipping box (metres) around point}
}
\description{
Clip grid to point plus error  # Clip Grids = 31
}
\keyword{internal}
