% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_creation.R
\name{CombineWatersheds}
\alias{CombineWatersheds}
\title{Combine Watersheds}
\usage{
CombineWatersheds(station, shedspoly_folder, dempoly_folder, out_folder)
}
\arguments{
\item{station}{station number in format "01XY001"}

\item{shedspoly_folder}{character, path to folder containing maximum
drainage basin shapefiles derived from a HydroSHEDS}

\item{dempoly_folder}{character, path to folder containing drainage basin
shapefiles derived from DEM}

\item{out_folder}{character, path to folder containing output files}
}
\description{
Combine DEM-derived and HydroBASINS-derived watershed
}
