% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDED_tools.R
\name{CompileDEM}
\alias{CompileDEM}
\title{Create DEM mosaic}
\usage{
CompileDEM(DEM, DEM.dir, output.dir, force.50k = F, product = "CDED")
}
\arguments{
\item{DEM}{character string, name of NTS tile (e.g. '075C' or '014D03') or USGS NED 1arc-sec tile
(e.g. 'n52e104'). If specified as
a 250k NTS tile (4 characters), a 250k NTS tile will be downloaded. If specified as a 50k NTS tile (6 character)
then a 50k NTS tile will be downloaded.}

\item{DEM.dir}{character string, path to directory where DEM tiles are stored. This function creates a nested
file hierarchy that matches that of the FTP site.}

\item{output.dir}{character string specifying where output file should be saved}

\item{force.50k}{logical, returns 50k tiles instead of 250k tiles even if the tiles are specified as 250k tiles.
e.g. using "075C" would use 075C01, 075C02, 075C03 and so on. Ignored if using 'NED' as data product.}

\item{product}{character, which DEM product to use.  One of ('CDED', 'CDEM', 'NED')}
}
\value{
name of output DEM
}
\description{
Downloads a set of DEM tiles and moasaicks them into one SAGA grid file.
}
\examples{
\dontrun{
CompileDEM("035D", "C:/temp", "C:/temp", "C:/temp/NTSmosaic.sdat", force.50k=T)
}
}
