% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HYDROBASINS_Rscript.R
\name{DelineateBasin}
\alias{DelineateBasin}
\title{Create basin delineation from DEM and HydroBASINS}
\usage{
DelineateBasin(station, con, outdir, HYBAS, DEM.path, DEM.index, DEM.source,
  saga.env, NCA, interactive = F, pourpoints, default.to.HYB = F,
  projected.CRS = "+proj=aea +lat_1=50 +lat_2=70 +lat_0=40 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs",
  ...)
}
\arguments{
\item{station}{Either a SpatialPointsDataFrame or the path to a Hydat-exported table with desired points
or a character vector of station names (in this case a connection must be specified)}

\item{con}{(optional) A open database connection.  Used if station is a character vector of station names}

\item{outdir}{(optional) A directory to which the output shapefile and report will be saved.  If not specified,
the function will return an R object without saving data to disk.}

\item{DEM.index}{index shapefile (optional)}

\item{DEM.source}{either}

\item{NCA}{(optional) Either a character string path or a SpatialPolygonsDataFrame.  This is a polygon
of the non-contributing areas.  If provided, calculates effective drainage area}

\item{pourpoints}{spatial points data frame with}
}
\value{
A list of SpatialPolygonDataFrames
}
\description{
Gives information about a station (upstream area, upstream stations)
}
