% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEM_tools.R
\name{DownloadAndUnzip}
\alias{DownloadAndUnzip}
\title{Download and unzip a file}
\usage{
DownloadAndUnzip(url, destfile, exdir, rmzip = T)
}
\arguments{
\item{url}{character url path}

\item{destfile}{character, filepath of output zipfile}

\item{exdir}{character,  the directory to which files are extracted}

\item{rmzip}{logical, whether or not to remove zipfile after extraction.
Defaults to TRUE.}
}
\description{
Downloads a zipfile from a URL and unzips it to a target
directory.
}
\details{
Runs as a subroutine of \link{DownloadSingleDEM}
}
