% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEM_tools.R
\name{DownloadMultipleDEM}
\alias{DownloadMultipleDEM}
\title{Download Multiple DEM files}
\usage{
DownloadMultipleDEM(DEM, DEM_dir, force50k = F, product = "CDED")
}
\arguments{
\item{DEM}{character string, name of DEM tile (e.g. '075C' or '014D03'). If
specified as
a 250k tile (4 characters), a 250k tile will be downloaded. If specified as a
50k tile (6 character)
then a 50k tile will be downloaded.}

\item{DEM_dir}{character string, path to directory where DEM tiles are stored.
This function creates a nested
file hierarchy that matches that of the FTP site.}

\item{force50k}{logical, returns 50k tiles instead of 250k tiles even if the
tiles are specified as 250k tiles.
e.g. using "075C" would use 075C01, 075C02, 075C03 and so on.}

\item{product}{either 'CDEM' or 'CDED'. Which dataset to download}
}
\description{
a wrapper for \link{DownloadSingleDEM} that allows a vector of
NTS sheets
or USGS DEM tile names to be specified
}
