% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEM_tools.R
\name{DownloadSingleDEM}
\alias{DownloadSingleDEM}
\title{Download DEM Sheet}
\usage{
DownloadSingleDEM(DEM_id, DEM_dir, replace = F, product = "CDED")
}
\arguments{
\item{DEM_dir}{character string, path to directory where DEM tiles are stored.
This function creates a nested file hierarchy that matches that of the FTP site}

\item{replace}{logical, degaults to FALSE, whether or not files should be
re-downloaded if they already exist in the DEM_dir directory}

\item{product}{character string specifying which data product to get (one of
'CDEM', 'CDED', 'CDSM', or 'NED')}

\item{NTS}{character string, name of NTS tile (e.g. '075C' or '014D03') or
USGS NED tile. If specified as a 250k tile (4 characters), a 250k tile will
 be downloaded. If specified as a 50k tile (6 character) then a 50k tile will
 be downloaded.}
}
\value{
a vector of file paths to downloaded DEM files
}
\description{
downloads DEM tiles from
"http://ftp.geogratis.gc.ca/pub/nrcan_rncan/elevation" and unzips
them to a directory
}
\details{
CDED and CDEM are from NRCAN. NED uses USGS National Elevation Dataset.
}
