% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ExpandBBox}
\alias{ExpandBBox}
\title{Expand Bounding box}
\usage{
ExpandBBox(geom1, tol)
}
\arguments{
\item{geom1}{an R Spatial* object}

\item{tol}{numeric, distance in kilometers to buffer}
}
\value{
a bounding box R object
}
\description{
Increases the size of a bounding box by a specified
}
\details{
This is used with \code{link[rcanvec]{nts.bbox}} in order to get all NTS tiles that
are within the buffer distance of the object
}
\keyword{internal}
