% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonames_db.R
\name{FilterKMLGeometry}
\alias{FilterKMLGeometry}
\title{Convert KML to shapefile, retaining only one geometry type}
\usage{
FilterKMLGeometry(kml.file, geom, try_alternate = T)
}
\arguments{
\item{kml.file}{character path to kml file}

\item{geom}{character string, which feature type do you want to retain? options
are one of ('point', 'line', 'polygon').}

\item{try_alternate}{logical, if no features of the desired type are found,
whether or not to try loading a different feature type.}
}
\value{
character, path to newly created shapefile
}
