% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HYDROBASINS_Rscript.R
\name{FindNearestRiverSegment}
\alias{FindNearestRiverSegment}
\title{Find nearest river segment}
\usage{
FindNearestRiverSegment(spatialPoint, riverLines, HYBAS)
}
\arguments{
\item{spatialPoint}{A SpatialPointsDataFrame of the desired station}

\item{riverLines}{A SpatialLinesDataFrame of the HydroSHEDS river network}
}
\value{
dataframe of nearest lines (attribute table of matches)
}
\description{
Find the nearest line segment (river) to a point (station).  Spatial objects should be in an
appropriate coordinate system (i.e. not lat-long) otherwise they will be projected and this adds to processing time.
If CRS does not match between the stations and the rivers they will both be projected into Canada Albers Equal
Area conic.
}
