% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HYDROBASINS_Rscript.R
\name{FindUpstreamStations}
\alias{FindUpstreamStations}
\title{Find Upstream Stations}
\usage{
FindUpstreamStations(HYBAS, HYBAS.ID, con, all = T, ...)
}
\arguments{
\item{HYBAS:}{a HYBAS spatial polygon object}

\item{HYBAS.ID:}{The HYBAS.ID of the sub-basin from which to go upstream}
}
\value{
A character vector of station IDs
}
\description{
Finds all HYDAT monitoring stations upstream of given HYBAS sub-basin
}
\keyword{internal}
