% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGA_Tools.R
\name{gdal_warp2SAGA}
\alias{gdal_warp2SAGA}
\title{Convert GRID to SAGA SGRD}
\usage{
gdal_warp2SAGA(srcfile, dstfile, output_crs, srcnodata = -32768,
  dstnodata = -32767, s_srs = "EPSG:4326")
}
\arguments{
\item{srcfile}{path to GRID folder (no trailing path separators)}

\item{dstfile}{(optional) path to output SAGA grid file.  If not specified,
saves to same directory as input.  Note that the file name must have the
"sdat" file extension (not sgrid)}

\item{output_crs}{character string specifying the coordinate system of the
output grid.  Examples include: "+proj=aea +lat_1=50 +lat_2=70 +lat_0=40
+lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"}

\item{s_srs}{source coordinate system.  Defaults to EPSG:4326 (WGS 84)}
}
\description{
Convertes a GRID DEM into SAGA .sgrd/.sdat format.  Also
converts the CRS into a projected system.
}
\keyword{internal}
