% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{get_downstream}
\alias{get_downstream}
\title{Get downstream stations}
\usage{
get_downstream(station, table)
}
\arguments{
\item{station}{target station for which a list of downstream
stations is desired}

\item{table}{a data.table describing upstream / downstream relationships between
stations.}
}
\value{
character vector of stations upstream of target station
}
\description{
return a list of stations that are downstream of a target station
}
\examples{
r <- initialize_matrix_from_stations(c('stn1', 'stn2', 'stn3'))
set_upstream('stn1', r, 'stn2')
get_upstream('stn1', r)

 @export
}
