% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{get_upstream}
\alias{get_upstream}
\title{Get upstream stations}
\usage{
get_upstream(station, table)
}
\arguments{
\item{station}{target station for which a list of upstream stations is desired}

\item{table}{a data.table describing upstream / downstream relationships between
stations.}
}
\value{
Does not return a value, the table is modified in-place.
}
\description{
return a list of stations that are upstream of a target station
}
\examples{
r <- initialize_matrix_from_stations(c('stn1', 'stn2', 'stn3'))
set_upstream('stn1', r, 'stn2')
get_upstream('stn1', r)

}
