% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEM_tools.R
\name{GetNTSDEMTilePath}
\alias{GetNTSDEMTilePath}
\title{Get FTP path for NTS DEM tile}
\usage{
GetNTSDEMTilePath(NTS, product = "CDED")
}
\arguments{
\item{NTS}{character, NTS tile name in format DDDLDD or DDDL (d=digit,
L=letter) e.g 075M05 or 014K}

\item{product}{character string specifying which data product to get (one of
'CDEM', 'CDED', 'CDSM')}
}
\value{
character, FTP path to desired NTS tile
}
