% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HYDROBASINS_Rscript.R
\name{getPourPoint}
\alias{getPourPoint}
\title{Move station to pourpoint}
\usage{
getPourPoint(point, pourpoint)
}
\arguments{
\item{point}{a SpatialPointsDataFrame representing one station. Must have a column named 'station_number'}

\item{pourpoint}{a SpatialPointsDataFrame of pourpoints. Must have a column named 'station_number'}
}
\value{
A SpatialPointsDataFrame with data identical to the original point,
and the new geometry corresponding to the pour point. If no matching pour point is found,
the original station point is returned
}
\description{
changes the location of a station (point) to that of its
pourpoint using the 'station_number' column as a shared ID.
}
