% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{GetProj4}
\alias{GetProj4}
\title{Proj4 string from abbreviation}
\usage{
GetProj4(x)
}
\arguments{
\item{x}{character string of a coordinate system. See options in details below}
}
\value{
proj4 string corresponding to x
}
\description{
A helper function to easily produce proj4 strings without introducting global variables
and without cluttering the main body of the code with long proj4 strings.
}
\details{
"WGS84" = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
 "AlbersEqualAreaConic" = "+proj=aea +lat_1=50 +lat_2=70 +lat_0=40
 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
 ... more to come
}
\keyword{internal}
