% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{GetTilePathsHS}
\alias{GetTilePathsHS}
\title{Get file paths of hydrosheds tiles}
\usage{
GetTilePathsHS(names, DEM_dir)
}
\arguments{
\item{names}{names of hydrosheds tiles (e.g. 'n45w060')}

\item{DEM_dir}{character path pointing to directory containing hydrosheds
tiles}
}
\value{
character, list of file paths for each of the tiles specified
in the names parameter
}
