% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{GraticuleIndices}
\alias{GraticuleIndices}
\title{Find degree-graticule tiles within distance}
\usage{
GraticuleIndices(long, lat, tol, resolution = 1, ...)
}
\description{
Find degree-graticule tiles within distance
}
