% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGA_Tools.R
\name{Grid2PolyRS}
\alias{Grid2PolyRS}
\title{Grid to polygon}
\usage{
Grid2PolyRS(in.GRD, out.poly, saga.env, CLASS_ID = 100, verbose = F)
}
\arguments{
\item{in.GRD}{character string path to SAGA grid file.  The grid should}

\item{out.poly}{character string path to output polygon shapefile}

\item{saga.env}{A SAGA geoprocessing object.  Suggested version is 2.2.2.}

\item{CLASS_ID}{numeric raster value that designates which grid cells to convert to polygon (cells with other values will be ignored)}
}
\description{
Creates a polygon from a grid ('shapes_grid' module 6: Vectorising Grid Classes)
}
