% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_creation.R
\name{HYBASBasinLimits}
\alias{HYBASBasinLimits}
\title{Create containing HydroBASINS}
\usage{
HYBASBasinLimits(df, HYBAS, code, outdir, hyb.id = "HYBAS_ID",
  df.id = "Station")
}
\arguments{
\item{df}{data.frame with exactly 1 row containing station names and the
HYBAS_ID of the HydroBASINS polygon that contains each station.}

\item{HYBAS}{SpatialPolygonsDataFrame of hydrobasins polygons}

\item{code}{character, classification code describing location of station
within cell and on stream}

\item{outdir}{character, directory in which to save output shapefile}

\item{hyb.id}{character, name of column containing ID of HydroBASINS sub-basin
that contains the station}

\item{stn.id}{character, name of column containing the station number}
}
\description{
Uses a classification code to build the maximum allowable
 drainage basin for a station using HydroBASINS polygons.
}
\details{
uses the following classification system
P: containing sub-basin has 2 input sub-basins
  Pcb Waterbody continues upstream and station on main branch below confluence
  Pct Waterbody continues upstream and station on tributary branch above confluence
  Pcm Waterbody continues upstream and station on main branch above confluence
  Ps Waterbody terminates upstream within sub-basin

I: containing sub-basin has 1 input sub-basins
  Ic Waterbody continues upstream
  Is Waterbody terminates upstream within sub-basin

T: containing sub-basin has 0 input sub-basins

The output shapefile contains the containing hydrobasin.
}
