% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_creation.R
\name{HYBASBasinLimits_Lake}
\alias{HYBASBasinLimits_Lake}
\title{Create HydroBASINS catchment for a hydrometric station on a lake}
\usage{
HYBASBasinLimits_Lake(station_point, HYBAS, lakes_folder, output_folder, code,
  stn_num_col = "station_number", stn_name_col = "station_name")
}
\arguments{
\item{station_point}{a Spatialpointsdataframe of the hydrometric station}

\item{output_folder}{character string, file path indicating where to save the
final output shapefile}

\item{code}{classification code describing station location. relative to
waterbodies and containing hydrobasins polygon}

\item{stn_num_col}{character string of the column name in station_point data
frame that contains the station name, defaults to "station_number"}

\item{stn_name_col}{character string of the column name in station_point data
frame that contains the station name, defaults to "station_name"}
}
\value{
A list containing two spatialpolygons data frames (the lake and
the catchment area)
}
\description{
Create Hydrobasins upstream boundaries for a hydrometric station
on a lake using the HydroBASINS data product
}
\details{
This function uses the following classification system for the
code parameter, which describes the location of the station relative to
the containing hydrobasin and the waterbody with which it is associated:

P: containing sub-basin has 2 input sub-basins
  Pcb Waterbody continues upstream and station on main branch below confluence
  Pct Waterbody continues upstream and station on tributary branch above confluence
  Pcm Waterbody continues upstream and station on main branch above confluence
  Ps Waterbody terminates upstream within sub-basin

I: containing sub-basin has 1 input sub-basins
  Ic Waterbody continues upstream
  Is Waterbody terminates upstream within sub-basin

T: containing sub-basin has 0 input sub-basins
}
