% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{initialize_matrix_from_stations}
\alias{initialize_matrix_from_stations}
\title{Create upstream matrix from station names}
\usage{
initialize_matrix_from_stations(stations, diagnl = 9)
}
\arguments{
\item{stations}{character vector of station names}

\item{diagnl}{integer used along the diagonal of the matrix to indicate
station identity (neither upstream nor downstream)}
}
\value{
an [n x n+1] data.table
}
\description{
return a list of stations that are downstream of a target station
}
\examples{
r <- initialize_matrix_from_stations(c('stn1', 'stn2', 'stn3'))

 @export
}
