% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{InterpretShapefile}
\alias{InterpretShapefile}
\title{Shapefile Helper}
\usage{
InterpretShapefile(x)
}
\arguments{
\item{x}{either an R spatial object or a character string specifying a shapefile path}
}
\description{
allows for shapefiles to be passed as character strings or
as R spatial objects in other functions
}
\keyword{internal}
