% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_creation.R
\name{Lake_Station}
\alias{Lake_Station}
\title{Make upstream basin boundaries for a station on a lake}
\usage{
Lake_Station(station_point, HYBAS, lakes_folder, DEM_path, DEM_source,
  outdir_DEM, outdir_HYBAS, saga_env, code, stn_num_col = "station_number",
  stn_name_col = "station_name")
}
\arguments{
\item{station_point}{a SpatialPointsDataFrame corresponding to a hydrometric station.
 Must be in same coordinates system as DEM and coordinate system
must be projected (not lat/long).  Must have longitude and latitude attributes.}

\item{HYBAS}{}

\item{lakes_folder}{}

\item{DEM_path}{One of:
(1) a file path to a directory containing DEM files either in the the format
 n%%w0%%_con_grid.sgrd (if DEM_source is 'SHEDS') or a directory to which DEM
 tiles will be downloaded (if DEM_source='CDEM', 'CDED' etc.),
(2) a file path to a dem file in SAGA format.  The DEM should be in a
projected coordinate system and the coordinate system should match that of
the point (e.g. Canada Albers Conformal Conic)}

\item{DEM_source}{character, one of: c('CDED', 'NED', 'CDEM, CDSM', 'SHEDS').
Ignored if a DEM file is supplied to DEM_path}

\item{outdir_DEM}{Directory to output final hydrobasins-derived delineation}

\item{outdir_HYBAS}{Directory to output final upslope (DEM) delineation}

\item{saga_env}{A SAGA geoprocessing object.  Suggested version is 2.2.2.}

\item{code}{character, classification code describing location of station
within cell and on stream}

\item{stn_num_col}{character string of the column name in station_point data
frame that contains the station name, defaults to "station_number"}

\item{stn_name_col}{character string of the column name in station_point data
frame that contains the station name, defaults to "station_name"}
}
\value{
does not return an R object but creates two basin delineation files,
one based on HydroBASINS and one basin on the DEM
}
\description{
This function acts a wrapper for two other functions
}
\details{
This function uses the following classification system for the
code parameter, which describes the location of the station relative to
the containing hydrobasin and the waterbody with which it is associated:

P: containing sub-basin has 2 input sub-basins
  Pcb Waterbody continues upstream and station on main branch below confluence
  Pct Waterbody continues upstream and station on tributary branch above confluence
  Pcm Waterbody continues upstream and station on main branch above confluence
  Ps Waterbody terminates upstream within sub-basin

I: containing sub-basin has 1 input sub-basins
  Ic Waterbody continues upstream
  Is Waterbody terminates upstream within sub-basin

T: containing sub-basin has 0 input sub-basins
}
