% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_creation.R
\name{LakeDEM}
\alias{LakeDEM}
\title{Calculate basin delineation for a hydrometric station on a lake using a DEM}
\usage{
LakeDEM(station_point, lake_poly, lake_hybas_limit, DEM_path, outdir, saga_env,
  DEM_source = "SHEDS", stn_name_col = "station_name",
  stn_num_col = "station_number")
}
\arguments{
\item{station_point}{a SpatialPointsDataFrame corresponding to a hydrometric
station. Must be in same coordinates system as DEM and coordinate system
must be projected (not lat/long).  Must have longitude and latitude
attributes. Must also have station name and station number attributes.}

\item{lake_poly}{SpatialPolygonDataFrame of lake that the station is on}

\item{lake_hybas_limit}{SpatialPolygonDataFrame of HydroBASINS upstream
area delineation as produced by HYBASBasinLimits_Lake}

\item{DEM_path}{One of:
(1) a file path to a directory containing DEM files either in the the format
 n%%w0%%_con_grid.sgrd (if DEM_source is 'SHEDS') or a directory to which DEM
 tiles will be downloaded (if DEM_source='CDEM', 'CDED' etc.),
(2) a file path to a dem file in SAGA format.  The DEM should be in a
projected coordinate system and the coordinate system should match that of
the point (e.g. Canada Albers Conformal Conic)}

\item{outdir}{Directory to output final upslope shapefile}

\item{saga_env}{A SAGA geoprocessing object.  Suggested version is 2.2.2.}

\item{DEM_source}{character, one of: c('CDED', 'NED', 'CDEM, CDSM', 'SHEDS').
 Ignored if a DEM file is supplied to DEM_path

 @param stn_num_col character string of the column name in station_point data
frame that contains the station name, defaults to "station_number"}

\item{stn_name_col}{character string of the column name in station_point data
frame that contains the station name, defaults to "station_name"}
}
\value{
No returned object, but creates a basin delineation from the DEM
}
\description{
Create a drainage basin shapefile using a DEM
}
\keyword{internal}
