% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HYDROBASINS_Rscript.R
\name{LoadHybas}
\alias{LoadHybas}
\title{Load a HydroBASINS shapefile into R}
\usage{
LoadHybas(path, area = "na", level = 7)
}
\arguments{
\item{area}{which region to load. can be 'na' (north america), or 'ar'
(arctic)}

\item{level}{The desired Pfaffstetter level}

\item{directory}{the directory containing HydroBASIN shapefiles.
Searches recursively.}
}
\value{
A list containing a spatial object and an attribute table
}
\description{
For a directory containing HydroBASIN shapefiles , load the file
associated with the specified Pfaffstetter level and return an attribute table.
NOTE: If a directory is provided, file naming convention must match original
HYBAS (e.g. hybas_na_lev02_v1c.shp)
}
\examples{
 LoadHybas("~/HydroBASINS", 'ar', 3)

}
