% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGA_Tools.R
\name{MoisaicRS}
\alias{MoisaicRS}
\title{Mosaic}
\usage{
MoisaicRS(grids, out_grid, saga_env, xmin, xmax, ymin, ymax, cellsize,
  verbose = F)
}
\arguments{
\item{grids}{character vector of input grids}

\item{out_grid}{file path to output grid}

\item{saga_env}{A SAGA geoprocessing object.  Suggested version is 2.2.2.}

\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{cellsize}{cell size}
}
\description{
Builds mosaicked DEM using SAGA 'grid_tools' module 3
}
