% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{most_downstream}
\alias{most_downstream}
\title{Get a list of the most downstream stations}
\usage{
most_downstream(table, diagnl = 9)
}
\arguments{
\item{table}{a data.table describing upstream / downstream relationships between
stations.}

\item{diagnl}{}
}
\value{
character vector of stations upstream of target station
}
\description{
return a list of stations that are the most downstream in their
catchment area
}
\examples{
r <- initialize_matrix_from_stations(c('stn1', 'stn2', 'stn3'))
set_upstream('stn1', r, 'stn2')
get_upstream('stn1', r)

 @export
}
