% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{NearestHYBAS}
\alias{NearestHYBAS}
\title{#' Hybas to stationprefix}
\usage{
NearestHYBAS(station)
}
\arguments{
\item{station}{character string specifying a station name e.g. "02AB001"}
}
\description{
For a station prefix, returns the first 3 digits of the pfaffstetter code corresponding
to the nearest hydrobasins polygon. This allows the hydrobasins file to be subset using indexing before
running through searches by removing polygons that are neither upstream nor downstream of the station
}
\keyword{internal}
