% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{NEDcoverage}
\alias{NEDcoverage}
\title{Find all NED DEM tiles necessary to cover spatial object}
\usage{
NEDcoverage(geom1, tol)
}
\arguments{
\item{geom1}{Spatial* object for which a covering set of tiles is desired}

\item{tol}{distance to buffer geometry (geom1)}
}
\description{

}
