% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydat_extras.R
\name{ParseStationName}
\alias{ParseStationName}
\title{Extract feature name from HYDAT station name}
\usage{
ParseStationName(station_name)
}
\arguments{
\item{station_name}{character string; station name from Hydat station}
}
\value{
character string corresponding to the hydrographic feature associated with the station (i.e.
the name of a lake, river, stream etc)
}
\description{
returns a string that is compatible with the canadian geographic names databse
}
\examples{
ParseStationName("PEARSON CREEK NEAR PROCTER")
ParseStationName("SCHIAVON CREEK BELOW HIGHEST DIVERSION NEAR THRUMS")
ParseStationName("MADAWASKA (RIVIER) EN AVAL DU BARRAGE TEMISCOUATA")
}
