% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{populate_table_from_delineations}
\alias{populate_table_from_delineations}
\title{Upstream stations from basin delineations}
\usage{
populate_table_from_delineations(table, polygon_folder, station_pts,
  station_number = "station_number")
}
\arguments{
\item{table}{a data table created from initialize_matrix_from_stations. An
(n x n+1) data.table with first column named "ID" and all other columns
named according to station number. Row i and column i+1 should correspond to
the same station.}

\item{polygon_folder}{character, path to folder containing delineation shapefiles}

\item{station_pts}{spatialpointsdataframe of monitoring stations}

\item{station_number}{column name identifying the station number}
}
\value{
A character vector of station IDs
}
\description{
Use a folder of basin delineations to define upstream relations
and populate upstream relations matrix
}
\keyword{internal}
