% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydat_extras.R
\name{read.pt2}
\alias{read.pt2}
\title{Read in a pt2 file from EC Data explorer output table}
\usage{
read.pt2(x)
}
\arguments{
\item{x}{pt2 file from EC Data explorer export}
}
\value{
A data.frame with (hopefully) appropriate data types
}
\description{
Read in a pt2 file from EC Data explorer output table. Does some type
conversion to make the output the same as the read.tb0 and read.dbf functions for ECDE
}
\details{
Read *.pt2
}
