% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read.sgrd.header}
\alias{read.sgrd.header}
\title{Read SAGA Grid header}
\usage{
read.sgrd.header(file)
}
\arguments{
\item{file}{character string path to *.sgrd}
}
\value{
data frame with list of grid parameters
}
\description{
get info for SAGA grid
}
\keyword{internal}
