% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_stations.R
\name{read.tb0}
\alias{read.tb0}
\title{Read *.tb0}
\usage{
read.tb0(x)
}
\arguments{
\item{x}{tb0 file from EC Data explorer export}
}
\value{
A data.frame with (hopefully) appropriate data types
}
\description{
Read in a tb0 file from EC Data explorer output table. Does some type
conversion to make the output the same as the read.pt2 and read.dbf functions for ECDE
}
