% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{read_coverage_tree}
\alias{read_coverage_tree}
\title{Read a station coverage text file}
\usage{
read_coverage_tree(coverage_file)
}
\arguments{
\item{coverage_file}{path to a coverage textfile describing which stations
are upstream of each station}
}
\value{
a list whose names correspond to stations and whose entries correspond
to the stations upstream of the named station.
}
