% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{readsimple}
\alias{readsimple}
\title{Get upstream stations in tree}
\usage{
readsimple(filepath)
}
\arguments{
\item{filepath}{character string path to coverage file describing upstream/
downstream station relationships}
}
