% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{SameCRS}
\alias{SameCRS}
\title{Ensure that two Spatial* objects have the same CRS}
\usage{
SameCRS(spgeom1, spgeom2)
}
\arguments{
\item{spgeom1}{Spatial* object that will be evaluated and transformed if necessary}

\item{spgeom2}{Spatial* object that will not be transformed}
}
\description{
Checks whether or not two objects have the same spatial reference.  If not, one of them
is transformed to match the other.
}
