% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{set_downstream}
\alias{set_downstream}
\title{Set downstream station}
\usage{
set_downstream(station, table, downstream, set_to = 1)
}
\arguments{
\item{station}{name of a target station for which downstream stations are to
be defined.}

\item{table}{a data.table describing upstream / downstream relationships between
stations.}

\item{set_to}{toggle between setting upstream (1) or not upstream(0)}

\item{upstream}{a character vector of stations in table that are downstream of
the target station.}
}
\value{
Does not return a value, the table is modified in-place.
}
\description{
For a target station, sets a list of stations as 'downstream'
}
\examples{
r <- initialize_matrix_from_stations(c('stn1', 'stn2', 'stn3'))
set_downstream('stn1', r, 'stn2')

 @export
}
