% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{set_downstream_recursive}
\alias{set_downstream_recursive}
\title{Recursively set downstream stations}
\usage{
set_downstream_recursive(station, table, next_downstream, set_to = 1)
}
\arguments{
\item{station}{target station to set}

\item{table}{a data.table describing upstream / downstream relationships between
stations.}

\item{next_downstream}{the next station downstream of the target station}

\item{set_to}{toggle between setting upstream (1) or not upstream(0)}
}
\description{
recursively sets all stations downstream of a target station
}
\examples{
M  <- initialize_matrix_from_stations(c('stn1', 'stn2', 'stn3', 'stn4', 'stn5'))
set_upstream('stn1', M, c('stn2', 'stn3', 'stn4'))
set_upstream('stn2', M, 'stn3')
set_downstream_recursive('stn5', M, 'stn3')
get_downstream('stn5', M)
}
