% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{set_relative_position}
\alias{set_relative_position}
\title{Set relative position of station}
\usage{
set_relative_position(station, table, next_upstream, next_downstream)
}
\arguments{
\item{station}{target station for which upstream-downstream relations
are to be set}

\item{table}{a data.table describing upstream / downstream relationships between
stations.}

\item{next_upstream}{(optional) the next station upstream of the target
station. If omitted, no upstream stations will be set.}

\item{next_downstream}{(optional) the next station downstream of the target
station. If omitted, no upstream stations will be set.}
}
\description{
, given the next-upstream and next-downstream stations, set
all up- and downstream stations as appropriate.
}
\examples{
M  <- initialize_matrix_from_stations(c('stn1', 'stn2', 'stn3', 'stn4','stn5'))
set_upstream('stn1', M, c('stn2', 'stn3', 'stn4'))
set_upstream('stn2', M, c('stn3', 'stn4'))
set_upstream('stn3', M, 'stn4')
set_relative_position('stn5', M, next_upstream='stn3', next_downstream='stn2')
get_upstream('stn5', M)
get_downstream('stn5', M)
}
