% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{set_upstream}
\alias{set_upstream}
\title{Set upstream station}
\usage{
set_upstream(station, table, upstream, set_to = 1)
}
\arguments{
\item{station}{name of a target station for which upstream stations are to
be defined.}

\item{table}{a data.table describing upstream / downstream relationships between
stations.}

\item{upstream}{a character vector of stations in table that are upstream of
the target station.}

\item{set_to}{toggle between setting upstream (1) or not upstream(0)}
}
\value{
Does not return a value, the table is modified in-place.
}
\description{
For a target station, sets a list of stations as 'upstream'
}
\examples{
r <- initialize_matrix_from_stations(c('stn1', 'stn2', 'stn3'))
set_upstream('stn1', r, 'stn2')

 @export
}
