% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{set_upstream_recursive}
\alias{set_upstream_recursive}
\title{Recursively set upstream stations}
\usage{
set_upstream_recursive(station, table, next_upstream, set_to = 1)
}
\arguments{
\item{station}{target station to set}

\item{table}{a data.table describing upstream / downstream relationships between
stations.}

\item{next_upstream}{the next station upstream of the target station}

\item{set_to}{toggle between setting upstream (1) or not upstream(0)}
}
\description{
recursively sets all stations upstream of a target station
}
\examples{
M  <- initialize_matrix_from_stations(c('stn1', 'stn2', 'stn3', 'stn4'))
set_upstream('stn1', M, c('stn2', 'stn3'))
set_upstream_recursive('stn4', M, 'stn1')
}
