% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{SnapToNearest}
\alias{SnapToNearest}
\title{Snap point to nearest line}
\usage{
SnapToNearest(point, lines)
}
\arguments{
\item{point}{spatialPointsDataFrame}

\item{lines}{spatialLinesDataFrame or spatialLines}
}
\value{
a spatial point with the same data as the input point, but on one of
the lines in lines.
}
\description{
moves a point to the nearest point on a line
}
\keyword{internal}
