% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_stations.R
\name{SpatialCSV}
\alias{SpatialCSV}
\title{Create SpatialPointsDataFrame from csv}
\usage{
SpatialCSV(file, header_x, header_y, CRS, ID)
}
\arguments{
\item{file}{path to *.csv file}

\item{header_x}{character, column name containing x coordinate information}

\item{header_y}{character, column name containing y coordinate information}

\item{CRS}{(optional) character, proj4 string specifying the projection
information of the points. If missing, the default is WGS84}

\item{ID}{character, column name giving a unique ID for each entry. If Hydat 'station_number' entries are
available for every station, this is the recommended column to use.}
}
\value{
SpatialPointsDataFrame
}
\description{
Creates an R SpatialPointsDataFrame from a csv
}
