% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydat_extras.R
\name{SpatialHydat}
\alias{SpatialHydat}
\title{Create a SpatialPointsDataFrame from a Hydat database connection.}
\usage{
SpatialHydat(con, ...)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database or a character string pointing to
a SQLite database}

\item{...}{Optional character vector of stations to subset (passed to HYDAT::StationMetaData))}
}
\value{
A SpatialPointsDataFrame for the Hydat stations
}
\description{
Create a SpatialPointsDataFrame from the "STATIONS" Hydat table.
}
\details{
SpatialHydat
}
