% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_names.R
\name{StationNameGeometry}
\alias{StationNameGeometry}
\title{Geometry of waterbody from station name}
\usage{
StationNameGeometry(name)
}
\arguments{
\item{name}{name of waterbody}
}
\description{
Attempts to figure out the geometry of a waterbody (point, polygon,
line) based on its name
}
\keyword{internal}
