% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{TileIndex}
\alias{TileIndex}
\title{Get names of DEM tiles that overlap object}
\usage{
TileIndex(geom1, tileindex, tile.id.field)
}
\arguments{
\item{geom1}{An R Spatial* object}

\item{tileindex}{either an R SpatialPolygonsDataFrame of the DEM tile index, or a character
path pointing to such a shapefile}

\item{tile.id.field}{Name of column in tileindex that gives DEM sheet number}
}
\description{
Find all DEM tiles that intersect a spatial object
}
\keyword{internal}
