% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{tree_get_all_upstream}
\alias{tree_get_all_upstream}
\title{Get all upstream stations from a tree graph}
\usage{
tree_get_all_upstream(station, tree)
}
\arguments{
\item{station}{name of station}

\item{tree}{list object returned from read_coverage_tree()}
}
\value{
list of all stations upstream of the specified station
}
