% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{tree_get_upstream}
\alias{tree_get_upstream}
\title{Get upstream stations in tree}
\usage{
tree_get_upstream(station, tree)
}
\arguments{
\item{station}{name of station}

\item{tree}{list object returned from read_coverage_tree()}
}
\value{
list of stations immediately upstream of the specified station
}
