% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{tree_to_matrix}
\alias{tree_to_matrix}
\title{Convert station tree to matrix}
\usage{
tree_to_matrix(tree)
}
\arguments{
\item{tree}{list object returned from read_coverage_tree()}
}
\value{
Matrix of station upstream/downstream relationships
}
\description{
converts an upstream/downstream network map from a tree-like
structure to a matrix structure.
}
