% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstReam.R
\name{upstream_stations_from_delineation}
\alias{upstream_stations_from_delineation}
\title{Find Upstream Stations}
\usage{
upstream_stations_from_delineation(basin, stations, target_station,
  station_number = "station_number")
}
\arguments{
\item{basin}{Either a spatialpolygonsdataframe of the target basin, or the
path to a shapefile of the target basin.  If a path is provided, the function
will attempt to use the sf package to hasten the shapefile read}

\item{stations}{SpatialPointsDataFrame of all hydrometric stations}

\item{target_station}{character, station number corresponding to the basin
that is provided. If provided, prevents the target station from being included
in the results as upstream of itself}

\item{station_number}{character, name of column in stations SpatialPointsDataFrame
which corresponds to the station number}
}
\value{
A character vector of station IDs
}
\description{
Finds all monitoring stations within a target basin
}
\keyword{internal}
