% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEM_tools.R
\name{USGSTileName}
\alias{USGSTileName}
\title{USGS Tile name from coordinates}
\usage{
USGSTileName(lon, lat, fext = "", v = "2013")
}
\arguments{
\item{lon}{longitude in degrees as an integer}

\item{lat}{latitude in degrees as an integer}

\item{fext}{file extension to append to file name}
}
\value{
a character string of format "n00w000%s"
}
\description{
Returns the base file name of a USGS NED elevation tile that
covers a set of coordinates
}
