% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEM_tools.R
\name{USGSTileURL}
\alias{USGSTileURL}
\title{USGS Tile URL}
\usage{
USGSTileURL(lon, lat, name, test = TRUE)
}
\arguments{
\item{lon}{longitude in degrees as an integer}

\item{lat}{latitude in degrees as an integer}

\item{name}{(optional)}
}
\value{
file path
}
\description{
Returns the URL of a USGS NED elevation tile
}
