% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonames_db.R
\name{CGNDBDownloadAPI}
\alias{CGNDBDownloadAPI}
\title{Download data from CGNDB}
\usage{
CGNDBDownloadAPI(dstfile, terms, format)
}
\arguments{
\item{dstfile}{character string specifying the path to a destination file
to store query results}

\item{terms}{a list of named search terms, see details for more info}

\item{format}{Character string; one of c('json', 'html', 'csv', 'kml', 'gml',
'wkt'). Determines the format of the results returned by the API.}
}
\value{
file path to downloaded file from Canadian Geographic Names Database
}
\description{
Generates a query and downloads the results
}
