% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeonamesAPI.R
\name{CGNDBQueryAPI}
\alias{CGNDBQueryAPI}
\title{Construct search query for CGNDB API}
\usage{
CGNDBQueryAPI(terms, format = NULL)
}
\arguments{
\item{terms}{a list() of named search terms, see details for more info}

\item{format}{Character string; one of c('json', 'html', 'csv', 'kml', 'gml', 'wkt').}
}
\value{
Character string; a URL-encoded search string
}
\description{
Construct search query for CGNDB API
}
\details{
More details about search terms can be found at
http://www.nrcan.gc.ca/earth-sciences/geography/place-names/tools-applications/9249
Important ones include q, bbox, lat, lon, radius, category, theme, province
Theme codes:
979  Water Feature
980 	Volcanic Feature
981 	Undersea and Maritime Feature
982 	Underground Feature
983 	Unclassified
984 	Terrain Feature
985 	Populated Place
986 	Ice and Snow Feature
987 	Feature Associated with Vegetation
988 	Constructed Feature
989 	Administrative Area
}
\examples{
QueryGeogratisAPI()
QueryGeogratisAPI(list(q='Ottawa', theme=985), format='csv')
}
