% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGA_Tools.R
\name{ChannelNetworkRS}
\alias{ChannelNetworkRS}
\title{Build channel network using RSAGA}
\usage{
ChannelNetworkRS(in.DEM, init.grid, out.shape, upstream.threshold = 35000,
  saga.env, verbose = F)
}
\arguments{
\item{in.DEM}{character string pointing to the *.sgrd path of a (hydrologically appropriate) DEM}

\item{init.grid}{(optional) character string pointing to a flow accumulation *.sgrd file.  If not
supplied, it will be created.}

\item{out.shape}{character string naming a shapefile open for writing which will be the channel network}

\item{upstream.threshold}{integer, number of upstream cells required to initiate channel growth.
A smaller number will yield more channels and will snap to smaller streams. Larger numbers will produce
only major channel branches. For a 50k DEM, numbers between 20,000 and 50,000 are appropriate to reproduce the
main channel network without many small branches.}

\item{saga.env}{A SAGA geoprocessing object.  Suggested version is 2.2.2.
character string  Mosaic 'grid_tools' module 3}
}
\value{
character string path to channel network shapefile
}
\description{
Builds a channel network from a DEM.
}
