% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEM_tools.R
\name{CompileDEM}
\alias{CompileDEM}
\title{Create DEM mosaic}
\usage{
CompileDEM(DEM, DEM_dir, output_dir, force50k = F, product = "CDED")
}
\arguments{
\item{DEM}{character vector, names of NTS tiles (e.g. '075C' or '014D03') or
USGS NED 1arc-sec tiles (e.g. 'n52e104'). If specified as a 250k NTS tile
(4 characters), a 250k NTS tile will be downloaded. If  specified as a 50k
NTS tile (6 character) then a 50k NTS tile will be downloaded.}

\item{DEM_dir}{character string, path to directory where DEM tiles are stored.
This function creates a nested file hierarchy that matches that of the
FTP site.}

\item{output_dir}{character string specifying where output file should be saved}

\item{force50k}{logical, returns 50k tiles instead of 250k tiles even if the
tiles are specified as 250k tiles. e.g. using "075C" would use 075C01, 075C02,
 075C03 and so on. Ignored if using 'NED' as data product.}

\item{product}{character, which DEM product to use.  One of ('CDED', 'CDEM',
'NED')}
}
\value{
name of output DEM
}
\description{
Downloads a set of DEM tiles and moasaicks them into one SAGA
grid file.
}
\examples{

CompileDEM("035D", "~", "~", "~/NTSmosaic.sdat", force50k = T)

}
