% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_creation.R
\name{HYBASBasinLimits}
\alias{HYBASBasinLimits}
\title{Create HydroBASINS delineation}
\usage{
HYBASBasinLimits(df, HYBAS, code, outdir, hyb_id = "HYBAS_ID",
  stn_id = "station_number")
}
\arguments{
\item{df}{data.frame with exactly 1 row containing station names and the
HYBAS_ID of the HydroBASINS polygon that contains each station.}

\item{HYBAS}{SpatialPolygonsDataFrame of hydrobasins polygons}

\item{code}{character, classification code describing location of station
within cell and on stream}

\item{outdir}{character, directory in which to save output shapefile}

\item{hyb_id}{character, name of column containing ID of HydroBASINS sub-basin
that contains the station}

\item{stn_id}{character, name of column containing the station number}
}
\value{
character, path to created shapefile
}
\description{
Uses a classification code to build an approximate
 drainage basin for a hydrometric station using HydroBASINS polygons.
}
\details{
This function uses the following classification system for the
code parameter, which describes the location of the station relative to
the containing hydrobasin and the waterbody with which it is associated:

P: containing sub-basin has 2 input sub-basins
  Pcb Waterbody continues upstream and station on main branch below confluence
  Pct Waterbody continues upstream and station on tributary branch above confluence
  Pcm Waterbody continues upstream and station on main branch above confluence
  Ps Waterbody terminates upstream within sub-basin

I: containing sub-basin has 1 input sub-basins
  Ic Waterbody continues upstream
  Is Waterbody terminates upstream within sub-basin

T: containing sub-basin has 0 input sub-basins

The output shapefile contains the containing hydrobasin.
}
