% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{HydroMosaic}
\alias{HydroMosaic}
\title{Find all hydroSHEDS tiles within a distance of a point}
\usage{
HydroMosaic(long, lat, tol, ...)
}
\arguments{
\item{tol}{Tolerance in m to consider}
}
\value{
A vector of grid names to load
}
\description{
Finds all HydroSHEDS DEM tiles within a specified tolerance of a point.  To be used
with 3 arc-second hydrologically conditioned DEM.
}
\keyword{internal}
