% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{HydroTile}
\alias{HydroTile}
\title{Get index of HydroSHEDS DEM tile}
\usage{
HydroTile(long, lat, fext = "")
}
\arguments{
\item{long}{numeric longitude}

\item{lat}{numeric latitude}

\item{fext}{file extension for output filename (needs leading '.')}
}
\description{
returns the name of a HYDROSHEDS tiled data file name based on the coordinates of a point
}
\keyword{internal}
