% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{LongitudeLength}
\alias{LongitudeLength}
\title{Calculate Length of Longitudinal Degree}
\usage{
LongitudeLength(lat, lat.m, lat.s)
}
\arguments{
\item{lat}{numeric vector - northing of location (degrees or decimal degrees)}

\item{lat.m}{(optional) numeric vector - northing of location, minutes (if lat is not specified in decimal degrees)}

\item{lat.s}{(optional) numeric vector - northing of location, seconds (if lat is not specified in decimal degrees)}
}
\value{
numeric vector of lengths  (measured in kilometres)
}
\description{
Calculates the length of one degree of longitude at a given latitude
}
