% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{outerRing}
\alias{outerRing}
\title{Extract largest ring}
\usage{
outerRing(poly)
}
\arguments{
\item{poly}{A SpatialPolygons* object with exactly feature/polygon. The feature may have multiple rings.}
}
\value{
SpatialPolygon
}
\description{
Returns the largest ring from a single polygon feature, retaining original ID value.
}
\details{
Finds the biggest ring of a SpatialPolygons* object and uses that ring to create
a new SpatialPolygons object with the same ID value. This can be used to clean up holes/overlap within polygons, or
to remove small polygons within a multipart polygon
}
