% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDED_tools.R
\name{OverlayDEM}
\alias{OverlayDEM}
\title{Create DEM mosaic to cover spatial object}
\usage{
OverlayDEM(geom1, tileindex, DEM.dir, output.dir, tol, product = "CDED",
  tile.id.field = "NTS_SNRC", ...)
}
\arguments{
\item{geom1}{An R Spatial* object to be 'buffered' with the DEM}

\item{tileindex}{spatialpointspolygon of DEM tile indices or path to shapefile of tile index}

\item{DEM.dir}{directory where DEM tiles are stored}

\item{output.dir}{where output DEM should be saved}

\item{tol}{(optional) by how many metres to buffer geom1}

\item{tile.id.field}{Name of column in tileindex that gives DEM sheet number}

\item{...}{optional arguments passed to \link{CompileDEM}}
}
\value{
name of output DEM
}
\description{
Covers a shape with a DEM grid
}
\details{
Builds a DEM from elevation dataset that either extends a fixed distance on either
side of a spatialobject (if a tol(erace) is specified) or intersects the DEM index (if
tileindex is specified)
}
