% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_names.R
\name{ParseStationName}
\alias{ParseStationName}
\title{Find waterbody name from HYDAT station name}
\usage{
ParseStationName(station_name)
}
\arguments{
\item{station_name}{character string; station name from Hydat station}
}
\value{
character string corresponding to the hydrographic feature associated
with the station (i.e. the name of a lake, river, stream etc)

a character string of the waterbody name
}
\description{
Given a NHS / Hydat hydrometric station name, attempts to isolate
the name of the waterbody on which the station resides.
}
\details{
The returned value can be used to query the canadian geographic
 names database to obtain a KML of the feature. The function relies on the
 regularity of the station naming pattern and may give unexpected results in
  certain circumstances such as:
 (*) when the branch of the river is specified
 (*) when there are multiple instances of parenthetical text in the name
 (*) when english and french are both used in the name
 .
}
\examples{
ParseStationName("PEARSON CREEK NEAR PROCTER")
ParseStationName("SCHIAVON CREEK BELOW HIGHEST DIVERSION NEAR THRUMS")
ParseStationName("MADAWASKA (RIVIER) EN AVAL DU BARRAGE TEMISCOUATA")

# some problem stations
ParseStationName("ROMAINE (RIVIERE) BELOW LAC LAVOIE")
ParseStationName("ARNAUD (PAYNE)(RIVIERE) EN AMONT DE LA RIVIERE HAMELIN-1" )

}
