% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{WhichZone}
\alias{WhichZone}
\title{Determine UTM zone from longitude}
\usage{
WhichZone(long)
}
\arguments{
\item{long}{numeric longitude in decimal degrees}
}
\value{
integer UTM zone number
}
\description{
Determines UTM zone from longitude
}
\details{
(floor((long + 180)/6) %% 60) + 1
}
\examples{
WhichZone(-125.7)

}
